Better separation of concern
A new method to sign messages directly was added to `SolanaWallet`. This
allows the user of the library to sign messages straight forward without
needing to serialize them.

Now the `SolanaClient` is no longer in charge of signing transactions.
Instead built the already signed transactions from the wallet. This way
we can use them anywhere and at any time. The solana client just has
functions to send and simulate transactions.

This in turn makes the `SolanaClient` class more general as it accepts
any `SignedTx` now and not just transfers. Some of these transactions
can be generated externally and signed with the new `signMessage()`
method of the `SolanaWallet`.