Make instructions and messages more generic
This makes building programs easier and cleaner

feat: Add support for the memo program

Now it is possible to pass arbitrary text to a tranfer transactions by
means of the spl-memo program.

To achieve this there was a major internal change in the code and the
way we had implemented the api. The api interface on the other hand had
minimal changes.