New api for building transactions.
Now we can build messages with the `Message` class by only specifying a list of instructions.

The class will automagically build the binary version of the message by using the information provided by each instruction.

It will collect all the accounts, sort them and build the required header, the binary list of of accounts as a `CompactArray` and of course the instructions themselves.

Within this commit as well, we removed the need for the `Serializable` interface as it's basically converting the information to their binary representation it makes more sense to just extend `Iterable<int>`.