Replace the implementation of `isOnCurve`
Our initial implementation of `isOnCurve` for ed25519 was wrong. Some public keys that were invalid were being considered valid erroneously. The derivation of the associated token address requires that the resulting address be off the curve. The algorithm keeps trying until it finds a _nonce_ (or in solana's source code called `bump_seed`) with which the resulting point is not on the curve. Given that our previous method of determining if a point was or not on the curve was giving false negatives, the generation of the associated token address (and it would affect the generation of any program specific account) was failing and thus the creation of the account on the system was failing as well.